+++ src/main/java/me/vasylkov/steamparser/parsing/model/ListingWithStickersMarkup.java
@@ -5,6 +5,7 @@ import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import me.vasylkov.steamparser.parsing.enums.ProfitableListingType;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -16,7 +17,7 @@ public class ListingWithStickersMarkup extends ProfitableListing
     private Double priceWithStickersMarkup;
     private Double stickersMarkupPercentage;
 
-    public ListingWithStickersMarkup(String listingId, String hashName, double price, List<Sticker> stickers, double floatValue, int pattern, String imgUrl, Double totalStickersPrice, Double priceWithStickersMarkup, Double stickersMarkupPercentage) {
+        public ListingWithStickersMarkup(String listingId, String hashName, double price, List<Sticker> stickers, BigDecimal floatValue, int pattern, String imgUrl, Double totalStickersPrice, Double priceWithStickersMarkup, Double stickersMarkupPercentage) {
         super(listingId, hashName, price, stickers, floatValue, pattern, imgUrl, ProfitableListingType.STICKERS);
         this.totalStickersPrice = totalStickersPrice;
         this.priceWithStickersMarkup = priceWithStickersMarkup;
